cleanup test files
